lazy loads modules at runtime.
An instance of Relution is set as prototype on the module exports. When a get property accessor
of an exposed member is read, a require statement is executed loading the actual implementation
code and the property is redefined to the resultant value.

For supporting diagnostics, the core module is an exception and must be loaded eagerly.